requirements.txt
django>=5.1
django-ninja>=1.0.0
